requirements.txt
# Core Framework
streamlit>=1.28.0
crewai>=0.36.0

# LangChain Ecosystem
langchain>=0.0.340
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Document Processing
PyPDF2>=3.0.1
pypdf>=3.16.2

# Vector Database
chromadb>=0.4.15

# Embeddings
ollama>=0.1.7

# Bioinformatics
biopython>=1.81

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Text Processing
regex>=2023.8.8

# Web and HTTP
requests>=2.31.0

# XML Processing
lxml>=4.9.3

# Machine Learning (for EfficientNet)
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
scikit-learn>=1.3.0

# Environment and Configuration
python-dotenv>=1.0.0

# Logging and Utilities
tqdm>=4.66.0

# Optional: For enhanced PDF processing
unstructured[pdf]>=0.10.0

# Optional: For better text splitting
sentence-transformers>=2.2.2

# Optional: For model optimization
accelerate>=0.24.0

# Optional: For GPU support (if using CUDA)
# torch-audio  # Uncomment if needed

# Development and Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Security and Encryption (optional for production)
cryptography>=41.0.0

# Monitoring and Observability (optional)
psutil>=5.9.0

# Optional: For better streamlit components
streamlit-extras>=0.3.5
streamlit-option-menu>=0.3.6

# For handling medical images
pydicom>=2.4.3
SimpleITK>=2.2.1

# For data validation
pydantic>=2.4.0

# For async operations
aiofiles>=23.2.0
asyncio-compat>=0.1.1

# For JSON handling improvements
orjson>=3.9.0

# For better error handling
tenacity>=8.2.0

# For caching
diskcache>=5.6.0

# For environment variable management
environs>=9.5.0